# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "create user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some Name"
                                            },
                                            "email": {
                                                "type": "email",
                                                "example": "demo@mail.ru"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "123456789"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully."
                                        },
                                        "User": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some Name"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "demo@mail.ru"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "user login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "email",
                                                "example": "demo@mail.ru"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "123456789"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "User": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some Name"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "demo@mail.ru"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "authorization": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL2xvZ2luI\n *                                                                                               iwiaWF0IjoxNjk5NDM5NDQyLCJleHAiOjE2OTk0NDMwNDIsIm5iZiI6MTY5OTQzOTQ0MiwianRpIjoiNGhsckNpaGJ\n *                                                                                               2QUltYTRwSyIsInN1YiI6IjEyIiwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.\n *                                                                                               7ZIPHuyYF8trjjEayQnjb1GaUQY9kn9uO23o2EWcAoU"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "user logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/forgetPassword": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "user forgetPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "email",
                                                "example": "demo@mail.ru"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Please open your email and click on url."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "refresh token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "User": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some Name"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "demo@mail.ru"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "example": "null"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "authorization": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL2xvZ2luI\n *                                                                                                iwiaWF0IjoxNjk5NDM5NDQyLCJleHAiOjE2OTk0NDMwNDIsIm5iZiI6MTY5OTQzOTQ0MiwianRpIjoiNGhsckNpaGJ\n *                                                                                                2QUltYTRwSyIsInN1YiI6IjEyIiwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.\n *                                                                                                7ZIPHuyYF8trjjEayQnjb1GaUQY9kn9uO23o2EWcAoU"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "user reset-password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "password": {
                                                "type": "string",
                                                "example": "Some password"
                                            },
                                            "confirmPassword": {
                                                "type": "string",
                                                "example": "Some Password"
                                            },
                                            "token": {
                                                "type": "string",
                                                "example": "aGFrb2JAbWFpbC5ydQ=="
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Your password changed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}